Update the button names to be more consistent and understandable.
We make sure that the popup alert is very similar to the one shown to
the user if they're in notification Alerts mode.  We rename "Stop" to
"Stop Timers" to make sure it's not confused with "stop the current
alarm sound".  Switch "Time to Work" to "Time FOR Work" to better
handle something like "Time for Downward-Facing Dog" or whatever.
Other similar changes.

We also switch the popup from an action sheet to an alert, to remove
similarity with the regular screen, and make it more clearly an
asynchronous modal situation like the user's more used to.