Make the button constraints better in tight spaces.
The 20-pt spacing between buttons makes the target area too small in a
constrained space, like in landscape mode, or an iPhone 4s.  Lower the
priority of the button spacing.  The new constraints look reasonable
on an iPhone 4s, an iPhone 6, and an iPad Air 2 in split-screen mode,
all in landscape or portrait, although the top text might stand to be
a bit larger on tablets.